requirements.txt
Django==3.2.4
django-assets==2.0
django-auth-ldap==2.4.0
django-braces==1.14.0
django-getenv>=1.3.2
django-honeypot==0.9.0
django-nose==1.4.7
django-post-office==3.5.3
django-simple-captcha==0.5.14
django-solo==1.1.5
django-widget-tweaks==1.4.8
eea.frame==0.5.5
gunicorn==20.1.0
jsonfield==3.1.0
psycopg2-binary==2.9.1
python-ldap==3.3.1
raven==6.10.0
requests==2.25.1
simplejson==3.17.2
six==1.16.0
unidecode==1.2.0
webassets==2.0
xlwt==1.3.0
XlsxWriter==1.4.3
xmltodict==0.12.0
